goods/requirements.txt
Django>=3.0.8
djangorestframework==3.11.0
psycopg2-binary==2.8.5
drf-yasg==1.17.1
django-sortedm2m==3.0.0
